+++ execution/federationtesting/gateway/main.go
@@ -25,9 +25,9 @@ func Handler(
 	httpClient *http.Client,
 	enableART bool,
 ) *Gateway {
-	upgrader := &ws.DefaultHTTPUpgrader
-	upgrader.Header = http.Header{}
-	// upgrader.Header.Add("Sec-Websocket-Protocol", "graphql-ws")
+	upgrader := &ws.HTTPUpgrader{
+		Header: http.Header{},
+	}
 
 	datasourceWatcher := datasourcePoller